Update ipwnder and ipwndfu for alloc8 stuff, prepare ipwndfu update/removal
fix ipwndfu reliability issue on nor dump for installing alloc8

also enable hacktivating on ipod4,1, closes #756

now that all of ipwnder can be used just fine for alloc8 installing, ipwndfu is no longer used for pwning at all.

the only use case left for ipwndfu is for alloc8 installation. it used to suck on apple silicon macs (random operation time outs all the time) but now that has been fixed as i tested.

now im working on getting ipwndfu working on python3, or even better, port/write the alloc8 installation process to c.

best case: ipwndfu gets completely replaced with a c implementation for alloc8 installation.

moderate case: ipwndfu is ported over to python3. i've tried this for a bit but i have moved on to working on a c port for now.

meh case: leave ipwndfu as is. pyenv/python2 is still used and persists along with the build times and possible issues. but at least it works ig