fix bootstraps
freeze now has patcyh, again. for the 64-bit 7.0-8.2 case (05f7da6 or #797), just rm the patcyh files and modify dpkg available and status as needed

additionally, with this commit 6.1.3 and 8.4.1 bundles are now removed in favor of using powdersn0w's dynamic patchers. this reduces the firmwarebundles to just 4.1 and lower, 4.2.1 for iphone 3g and touch 2, and all of 3gs release versions.